Split e2e suites into separate files
